+++ cob_twist_controller/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package cob_twist_controller
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.7.5 (2019-05-20)
+------------------
+* Merge pull request `#206 <https://github.com/ipa320/cob_control/issues/206>`_ from fmessmer/improve_lookat_extension
+  improve lookat extension
+* implement lookat_pointing_frame
+* fix lookat_chain offset, allow KinematicExtension to fail, param-config converter functions
+* Contributors: Felix Messmer, fmessmer
+
 0.7.4 (2019-04-05)
 ------------------